+++ CHANGELOG.md
@@ -25,6 +25,8 @@
   `SendContact`, `SendLocation` and `SendVenue` methods, use `ephemeralMessageParameters` parameter instead.
 - Enh #217: Make `text` parameter of `EditEphemeralMessageText` method optional.
 - Enh #214: Rework `.gitattributes` to explicitly allow only distribution files for export.
+- Bug #218: Register `PaidMediaLivePhoto` type in `PaidMediaValue` parse result value processor, without it
+  parsing of paid media with `live_photo` type failed.
 
 ## 0.21.2 July 15, 2026